refactor: simplify the logic around documentation generation
- remove all the unnecessary convoluted helpers
- remove the documentation prop from the extension interface
- generate documentation in the webserver and use basic caching mechanism
to avoid loading documentation from disk more than once
