+++ src/__tests__/extensions.test.ts
@@ -1,11 +1,12 @@
 import { isEmpty } from 'lodash'
 import { extensions } from '../../extensions'
+import { getExtensionDocumentation } from '../documentation'
 
 describe('Extensions', () => {
   describe('All extensions should have documentation (i.e. a README file in their root dir)', () => {
     test.each(extensions)('Check $key extension has documentation', (ext) => {
-      expect(ext).toHaveProperty('htmlDocs')
-      expect(isEmpty(ext.htmlDocs)).toBe(false)
+      const documentation = getExtensionDocumentation(ext.key)
+      expect(isEmpty(documentation)).toBe(false)
     })
   })
 })
