--- src/utils/index.ts
@@ -1,3 +1,0 @@
-export { getAppRootDir } from './getAppRootDir'
-export { getAllExtensionReadmes } from './getAllExtensionReadmes'
-export { withDocumentation } from './withDocumentation'
