--- src/utils/withDocumentation/index.ts
@@ -1,1 +1,0 @@
-export { withDocumentation } from './withDocumentation'
